Move ResolvedCall, ProcedureSignature (and friends) back to the compiler module
