+++ community/cypher/cypher-compiler-3.0/src/main/scala/org/neo4j/cypher/internal/compiler/v3_0/spi/ProcedureModeSupport.scala
@@ -19,8 +19,6 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_0.spi
 
-import org.neo4j.cypher.internal.frontend.v3_0.spi.{ProcedureAccessMode, ProcedureReadOnlyAccess, ProcedureReadWriteAccess}
-
 object ProcedureModeSupport {
   implicit class ProcedureModeSupportWrapper(mode: ProcedureAccessMode) {
     def callMode: ProcedureCallMode = mode match {
